requirements.txt
python-dotenv
flask
gunicorn
requests
